move select_parent into htsohm.py
