+++ htsohm/htsohm.py
@@ -1,10 +1,9 @@
 from math import sqrt
 
-from sqlalchemy.sql import func
+from sqlalchemy.sql import func, or_
 from sqlalchemy.orm.exc import FlushError
 
 from htsohm import config
-from htsohm.binning import select_parent
 from htsohm.db import session, Material, MutationStrength
 from htsohm.simulate import run_all_simulations
 from htsohm.material_files import write_seed_definition_files, write_child_definition_files
@@ -31,6 +30,57 @@ def evaluate_convergence(run_id):
     bin_counts = [i[0] for i in bin_counts]    # convert SQLAlchemy result to list
     variance = sqrt( sum([(i - (sum(bin_counts) / len(bin_counts)))**2 for i in bin_counts]) / len(bin_counts))
     return variance
+
+
+def select_parent(run_id, max_generation, generation_limit):
+    """Use bin-counts to preferentially select a list of rare parents.
+
+    Each bin contains some number of materials, and those bins with the fewers materials represent
+    the most rare structure-property combinations. These rare materials are preferred as parents
+    for new materials, because their children are most likely to display unique properties. This
+    function first calculates a `weight` for each bin, based on the number of constituent
+    materials. These weights affect the probability of selecting a parent from that bin. Once a bin
+    is selected, a parent is randomly-selected from those materials within that bin.
+    """
+
+    # Each bin is counted...
+    bins_and_counts = session \
+        .query(
+            func.count(Material.id),
+            Material.methane_loading_bin,
+            Material.surface_area_bin,
+            Material.void_fraction_bin
+        ) \
+        .filter(
+            Material.run_id == run_id,
+            or_(Material.retest_passed == True, Material.retest_passed == None),
+            Material.generation <= max_generation,
+            Material.generation_index < generation_limit,
+        ) \
+        .group_by(
+            Material.methane_loading_bin, Material.surface_area_bin, Material.void_fraction_bin
+        ).all()[1:]
+    bins = [{"ML" : i[1], "SA" : i[2], "VF" : i[3]} for i in bins_and_counts]
+    total = sum([i[0] for i in bins_and_counts])
+    # ...then assigned a weight.
+    weights = [i[0] / float(total) for i in bins_and_counts]
+
+    parent_bin = np.random.choice(bins, p=weights)
+    parent_query = session \
+        .query(Material.id) \
+        .filter(
+            Material.run_id == run_id,
+            or_(Material.retest_passed == True, Material.retest_passed == None),
+            Material.methane_loading_bin == parent_bin["ML"],
+            Material.surface_area_bin == parent_bin["SA"],
+            Material.void_fraction_bin == parent_bin["VF"],
+            Material.generation <= max_generation,
+            Material.generation_index < generation_limit,
+        ).all()
+    potential_parents = [i[0] for i in parent_query]
+
+    return int(np.random.choice(potential_parents))
+
 
 def mutate(run_id, generation, parent):
     """Retrieve the latest mutation_strength for the parent, or calculate it if missing.
